// frontend/src/app/solicitudes-recursos/models/solicitud-arma.ts
export interface SolicitudArma {
  url: string;
  idSolicitudArma: string;
  coordXAsentamiento: number;
  coordYAsentamiento: number;
	coordXPuntoCaida: number;
  coordYPuntoCaida: number;
	alcanceMax: number;
	zonaSegAngulo: number;
  armaUrl: string;
  armaId: string;
  arma: string;
  solicitudUrl: string;
  solicitudId: string;
  solicitud: string;
}